Add files via upload
This is an interactive quiz app which provides an interactive and engaging platform for users to test their knowledge. By combining HTML for structure, CSS for styling, and Javascript for dynamic functionality. This project serves as a practical example of building a web-based application. Further enhancements can be made by adding more questions, improving styling, or integrating additional features for an enriched user experience.